separate class out
